Fix return type mismatch in match_content_examples function
- Change engagement_rate from decimal to numeric (proper PostgreSQL type)
- Change similarity from float to double precision (matches internal storage)
- Apply to both bootstrap (001) and migration (004) files
- Fixes 'return type mismatch in function declared to return record' error

+++ sql/004_fix_content_examples_filter.sql
@@ -22,12 +22,12 @@ RETURNS TABLE (
   platform text,
   content text,
   human_score integer,
-  engagement_rate decimal,
+  engagement_rate numeric,
   content_type text,
   creator text,
   hook_line text,
   tags text[],
-  similarity float
+  similarity double precision
 )
 LANGUAGE sql STABLE
 AS $$
